migrate: move vamcHealthServicesListing to products directory
- Migrated vamcHealthServicesListing product files to src/products/vamcHealthServicesListing/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved query.ts from src/data/queries/ to src/products/vamcHealthServicesListing/
- Moved test file and renamed vamcHealthServicesListing.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vamcHealthServicesListing/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/vamcHealthServicesListing/mock.json
- Updated all import statements across the codebase
- Fixed import paths in migrated files (formatted-type.ts, template.test.tsx)
- All TypeScript errors resolved
- All tests passing (3 query tests + 3 template tests)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and vamcHealthServicesListing pages load correctly
