+++ .github/CODEOWNERS
@@ -26,3 +26,4 @@ src/products/vetCenterLocationListing/ @department-of-veterans-affairs/ap-team @
 src/products/vetCenterOutstation/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
 src/products/vamcSystem/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
 src/products/healthCareLocalFacility/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
+src/products/vamcHealthServicesListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
