+++ src/products/vamcHealthServicesListing/query.ts
@@ -1,7 +1,7 @@
 import { QueryData, QueryFormatter, QueryParams } from 'next-drupal-query'
 import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
 import { NodeVamcHealthServicesListing } from '@/types/drupal/node'
-import { VamcHealthServicesListing } from '@/types/formatted/vamcHealthServicesListing'
+import { VamcHealthServicesListing } from '@/products/vamcHealthServicesListing/formatted-type'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 import { ExpandedStaticPropsContext } from '@/lib/drupal/staticProps'
 import {
